requirements.txt
alembic
black
SQLAlchemy
requests
python-dotenv
